requirements.txt
arcgis
dask[dataframe]
pandas
docx-mailmerge
Office365-REST-Python-Client
python-dotenv